don't use EEControlBits::FORMAT100 in inline editengine for kit mode
With multiple users the EditEngine for the first user is created during
load of the document, before ScModelObj::initializeForTiledRendering is
called, and that sets SetTextWysiwyg(true)

ScInputHandler::UpdateRefDevice has...
if ( bTextWysiwyg ... ) mpEditEngine->SetControlWord( ... | EEControlBits::FORMAT100 );

That FORMAT100 means that layout for such edit engines is done at a X/Y
Scaling of 1

The first users editengine doesn't get any of that set, because
TextWysiwyg is not true yet when it's created, and so matches the normal
document rendering, but later joiners have TextWysiwyg set, so trigger
this alternative path and so the mapmode doesn't match the other case
and the two users have different results for the width of text in an
active editengine.

As it turns out, the results without FORMAT100 are the preferred ones,
so make that explicit for kit.

Change-Id: I76a78be1fca4af84c493d32bdd43f968ba072452
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/164058
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>