tdf#99969: make sure to copy the chart source ranges to clipboard
... even when they are outside of the copied cell range. Otherwise,
it is  impossible to transfer the missing data when switching to
own data on paste.

The copy of the missing ranges avoids copying cell attributes, for
performance reasons, but also to avoid overwriting the attributes
of already copied cells. Otherwise, ScDrawLayer::CopyToClip would
need the bKeepScenarioFlags, or the CopyToClipContext used in the
caller ScDocument::CopyToClip, for consistent copy; or a method to
avoid overwriting already copied cells (this change simply copies
all chart ranges, withiut checking if they were copied already).

Change-Id: Id02e0c20517e7e8a17bb0a31d1b230196cda1a58
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/164294
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Signed-off-by: Xisco Fauli <xiscofauli@libreoffice.org>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/164397