cool#8465 sw lok: classify anchored images as complex selection
Regression from commit 7a8dc25defee31edbb75a2f8c35f92ee2d3f3a83 (sw lok:
simplify SwTransferable::isComplex(), 2021-02-23), in case as-char
imagse were part of a selection, we considered that complex, but at-char
was considered as simple, which is inconsistent.

This was not intentional, simply the rework to avoid copying the
selection to a temporary document lost this functionality.

Fix the problem by using CollectFrameAtNode() to find at-char images,
which tries to use the layout, so is not meant to be too slow.

An alternative would be sw::GetFlysAnchoredAt(), but that doesn't try to
use the layout, so avoid that.

(cherry picked from commit 1bca99617ad54d966625caadd71e52134c70ae44)

Change-Id: I647d0f3934f9553de69afbf2de059631e2b5619c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/164521
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>