exported SVGs: Stop bullets skipping animation
Previously, we used the incorrect format for bullet point IDs, leading
to them not being noticed by animation code, leading to them being
skipped during animation.

This meant that if you exported an SVG of an impress presentation, and
you had a bullet list animating to appear later, you would see the
bullets without the text from the start.

Change-Id: Ibce764c0843778dd09b108fb251ce606255afb90
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/164661
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Gökay ŞATIR <gokaysatir@collabora.com>