tdf#160436: fix glyph bounds calculation for vertical glyphs
It is unclear if LogicalFontInstance::GetGlyphBoundRect can be called
for both normal and rotated variants of the same glyph in the same font.
If yes, then the normal and vertical variants must be cached separately,
or possibly vertical variant can be not cached, but always calculated.
This problem already existed before, so this change doesn't introduce
a new issue.

Change-Id: I9b50ef340c9e38db7bef890165519aadc96d3ffa
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/165581
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
(cherry picked from commit dff57d2b4f5abd1b51ebfb0015339471f61e72f9)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/165843
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>