annot: moved more of Annotation and dependencies into svx
Moved the holder of annotations from SdPage to SvxPage, so that
the vector holding the annotations and accessors are on SvxPage
and adapted the code.

This also changes the type od most parameters on most methods
from sd::Annotation to sdr::annotation::Annotation and adapted
the code.

Moved AnnotationEnumeration into svx, as it was needed in svx
already.

Change-Id: Iab17aa881443f58adfb9158959db00ed24076279
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/166494
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
(cherry picked from commit a0a581ead18f030f59d203539706de0230746cae)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/168555
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>