annot: move TextAPI to svx and adapt the code
TextApiObject is derived from SvxUnoText and holds and handles the
text in the annotation. This functionality is changed to get rid
of sd module specifics and moved to svx module. This now allows to
move the most improtant parts of the Annotation object to svx.

Change-Id: I34e238616e7fac973e75e79bb60a8d093c493258
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/166497
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
Tested-by: Tomaž Vajngerl <quikee@gmail.com>
(cherry picked from commit 2791964914ba19b636ce49a60508bcab7af675cd)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/168559
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>